Moved DoubleReuseRandom to the utilities package.
